Don't write tyk.conf when running the tests
Also simplify overriding of default config values for the tests.

On top of that, set EnableAnalytics to true for all the tests. Setting
it to true just in a couple of tests and not setting it back when they
are done means the config will be different depending on the test order.
Moreover, it would break further once we enable parallel tests.