test: don't set up global analytics three times
It's already set in initialiseSystem, and the two other times we set it
up in tests it's with an equal store. Moreover, this is a global so it
ends in data races when 'go test -race' is used.

Since it's a global we can't use a per-test unique KeyPrefix, so just
give up and set it up in TestMain.

Also fix the stutter in the test name.