utils/ci-test: timeout at 30s
The default timeout is 10m. This means if our tests are flakey and time
out (for example because of redis pub/sub being flakey, which has hit
Matias and I in the past), we can get a red CI report in less than a
minute instead of in more than 10m.