Limit hot reloads to once per 1s, not 10s
It makes sense to have some limit, as we don't want to allow continuous
reloads which could spin the CPU and slow down the gateway.

On the other hand, 10s seems a bit excessive. 1s still keeps the cpu
from spinning, yet reduces the amount of time that a user has to wait
for a reload from 0-10s to 0-1s.

This means we can say that hot reloads should take effect within a
second, which means they take effect almost instantaneously.