Fix ApiUpdated test to properly run its reload
On one hand it was running a reload manually, which was missing the
point of the test. We want the redis event to fire it on its own.

On the other, not sending a reload tick would mean that if the event
indeed fires a reload, it would be queued but not executed, breaking
other tests that depend on an empty reload queue.