jsvm: add a timeout and use it to kill the VM
Unfortunately, the only way to stop/kill Otto is via a panic, which
makes this code more complicated than it should be.

Make the Timeout be 5 seconds for now. It's high enough that no sane
middleware would hit it, but low enough to not let goroutine leaks build
up easily.

Fixes #496.