api: rework to vastly simplify json logic
Each "handler" func would return raw JSON instead of an object to be
marshaled. This means that the logic to write a JSON response was
copy-pasted in dozens of places, including what to do when marshaling
fails.

Instead, have these return an interface{} and have the JSON be handled
in fewer places, and with a proper helper func.

This also lets us get rid of a lot of "success" logic clutter.