api: replace some erroneous 403 status codes
* If the DB failed, that's a 500 (internal server error)
* If the session/key was wrong/missing, that's a 400 (bad request)
* If something was not found, that's a 404 (not found)
* If a key/token expired, that's a 401 (unauthorized, not logged in)

Note that there is nothing to fix in the oauth2 module in this manner,
as status codes are handled entirely in the osin library. That library
only allows for one OK and one error status code, which are 200 and 403
(as configured in main.go). Which is correct, as in the oauth2 scenario
the user is always logged in.