api: switch to two reload workers instead of one
We had one worker which was executing the actual reloads. To be able to
have the "done" callback funcs attach to previously queued reloads, we
must add a second worker goroutine that keeps track of these without
data races nor active polling.

Now, all communication goes through the new goroutine, as it's the one
that manages the queue.

This also fixes the case where a reload that wasn't queued didn't block.
The proof is in TestHotReloadMany - that func now expects 25 callbacks
when it sends 25 reloads to the queue, when only one reload is actually
executed under the hood.