roundrobin: lock structure with a mutex
The host checker and load balancer are wonky if the gateway gets
many requests while the host checker marks a host as down.

This smells of a data race. Sure enough, making the committed changes to
the test and running it with -race yields:

	WARNING: DATA RACE
	Write at 0x00c420478398 by goroutine 129:
	  github.com/TykTechnologies/tyk.GetNextTarget()
	      /home/mvdan/go/src/github.com/TykTechnologies/tyk/api_definition.go:1116 +0x152
	  github.com/TykTechnologies/tyk.TestHostChecker.func3()
	      /home/mvdan/go/src/github.com/TykTechnologies/tyk/host_checker_test.go:116 +0x7d

	Previous write at 0x00c420478398 by goroutine 188:
	  github.com/TykTechnologies/tyk.GetNextTarget()
	      /home/mvdan/go/src/github.com/TykTechnologies/tyk/api_definition.go:1116 +0x152
	  github.com/TykTechnologies/tyk.GetNextTarget()
	      /home/mvdan/go/src/github.com/TykTechnologies/tyk/tyk_reverse_proxy_clone.go:127 +0x674
	  github.com/TykTechnologies/tyk.TestHostChecker()
	      /home/mvdan/go/src/github.com/TykTechnologies/tyk/host_checker_test.go:134 +0xa56
	  testing.tRunner()
	      /home/mvdan/tip/src/testing/testing.go:754 +0x16c

The issue is that we're writing and reading the RoundRobin struct while
other goroutines might be doing the same. Use a lock to guard its use.

Add a TODO for the future, since this can mean a contention bottleneck -
only one request can use the round robin at a time. For now, this is
a safe and acceptable change for backporting.

Lastly, remove the "pos > targetData.Len()-1" logic. Now that we guard
our use of RoundRobin with a lock, this is impossible. SetLen will
already have reset pos to 0 if pos was larger than the new length.

Updates #825.