revproxy: deduplicate GetNextTarget code
We had the same 8 lines of code three times.

When UseDiscoveryService is true, we did it either way, even if
EnableLoadBalancing was false - but replacing the target host list.

And we do it if EnableLoadBalancing is true, period.

Keep the same logic, but deduplicate the code. This will make changing
the behaviour when there is no live host easier, as we only need to fix
one call site instead of three.

For #836.