main: remove docopt error warning
Since 36775eb4, we've told docopt to exit on any user errors instead of
returning them. The only other kind of errors that can happen is if the
usage text is malformed, in which case something went really wrong and
we want to panic instead of printing a warning and moving on.