mw: remove useContext param from urlRewrite
It was always true in production code. And in the tests it was false.
This was presumably because of gorilla/context, which might misbehave
given how test requests don't have any context values.

Now that we use net/http request contexts, there's nothing to worry
about.

Most importantly, we want the test to run the same code that prod would.
And if we want to add tests that use $tyk_context-style variables, we
now can.