api: re-allow empty keyName path argument in GETs
When I removed all ".*" regexes from gorilla/mux, I thought the default
regex was "[^/]*" - i.e. that an empty value was accepted. But the
default value is actually "[^/]+", where an empty value doesn't match
the route.

This broke a few endpoints that used an empty keyName in GET requests to
list all the keys.

Modify one of the tests to check this.

Fixes #901.