hostcheck: fix test ticker flake
When running the tests, the host checker uses a ticker channel so that
the tests can control when the uptime checks happen without relying on
times and sleeps.

However, the host checker only blocked waiting for the tick after doing
a host check. In other words, the first host check would happen without
waiting for a tick.

This could lead to a test receiving two host check events when it
expected exactly one.

Move the blocking channel receive to not do that first host check at a
random time.

Fixes #919.