revproxy: return a 503 if all hosts are down
Instead of directing to any one of the hosts.

Due to the nature of how the proxy is implemented, the load balancing
happens in the director. At that stage, we can only change the target
URL, not directly write to the response.

Hence, we must start up a local http server that will simply write the
503 and an error message. Use sync.Once to make sure it's not started
multiple times, and reuse it.

Fixes #836.