mw: remove IsEnabledForSpec dups in ProcessRequest
In a middleware, ProcessRequest can only be called if IsEnabledForSpec
returned true. Thus, it makes little sense for middlewares to repeat or
do their "enabled" check in ProcessRequest.

This was the case for a few of them. Turns out that this was done
because some parts of the code, especially the tests, did not always
keep that promise.

Refactor the codebase to keep that promise and apply the
simplifications. Add a new helper func, mwList, to help us with that.
Also renamed appendMiddleware to make it more obvious that the actual
append doesn't always happen.